Drop support for flannel
The default was switched to calico over 2 months ago - and almost all
testing has been happening exclusively on calico. So remove the support
for the alternate CNI provider.

With this, we can remove cni plugins tarball as well. Calico pushes
the required plugins to the nodes when it is deployed.